Feature/system dashboards (#26)
* System dashboard property was added. Dashboard service logic was improved according system requirements.
Personal and User Performance dashboards visibility was updated to public.

* Incorrect placeholder was fixed

* Update dashboard position functionality was blocked for system dashboards
System property was added to dashboard DTO

+++ service/src/main/java/com/zebrunner/reporting/service/DashboardService.java
@@ -23,10 +23,12 @@ import static com.zebrunner.reporting.service.exception.ResourceNotFoundExceptio
 public class DashboardService {
 
     private static final String ERR_MSG_DASHBOARD_CAN_NOT_BE_FOUND = "Dashboard with id %s can not be found";
+    private static final String ERR_MSG_DASHBOARD_CAN_NOT_BE_FOUND_BY_ATTRIBUTE_ID = "Dashboard which contains an attribute with id '%d' can not be found";
     private static final String ERR_MSG_DASHBOARD_ALREADY_EXISTS = "Dashboard with such title already exists";
     private static final String ERR_MSG_UNEDITABLE_DASHBOARD_CANT_BE_ALTERED = "Uneditable dashboard can not be updated or deleted";
     private static final String ERR_MSG_DASHBOARD_ATTRIBUTES_ALREADY_IN_USE = "Attributes with keys (%s) already in use for dashboard with id '%d'";
     private static final String ERR_MSG_DASHBOARD_ATTRIBUTE_ALREADY_IN_USE = "Attribute with key '%s' already in use for dashboard with id '%d'";
+    private static final String ERR_MSG_CANNOT_PERSIST_ATTRIBUTE_TO_SYSTEM_DASHBOARD = "Can not create or update attribute";
 
     private final DashboardMapper dashboardMapper;
     private final UserPreferenceService userPreferenceService;
@@ -79,7 +81,8 @@ public class DashboardService {
     public Dashboard update(Dashboard updatedDashboard) {
         Dashboard dashboard = getDashboardById(updatedDashboard.getId());
         // only editable dashboard can be modified, throw exception otherwise
-        if (dashboard.isEditable()) {
+        boolean allowedToUpdate = dashboard.isEditable() && !dashboard.isSystem();
+        if (allowedToUpdate) {
             updatedDashboard.setEditable(true);
             dashboardMapper.updateDashboard(updatedDashboard);
 
@@ -105,7 +108,8 @@ public class DashboardService {
     public void removeById(Long id) {
         Dashboard dashboard = getDashboardById(id);
         // only editable dashboard can be deleted, throw exception otherwise
-        if (dashboard.isEditable()) {
+        boolean allowedToRemove = dashboard.isEditable() && !dashboard.isSystem();
+        if (allowedToRemove) {
             // reset dashboard preference first, then delete
             userPreferenceService.resetDefaultDashboardPreference(dashboard.getTitle());
             dashboardMapper.deleteDashboardById(id);
@@ -128,9 +132,14 @@ public class DashboardService {
 
     @Transactional
     public List<Widget> updateDashboardWidgets(Long dashboardId, List<Widget> widgets) {
-        return widgets.stream()
-                      .map(widget -> updateDashboardWidget(dashboardId, widget))
-                      .collect(Collectors.toList());
+        Dashboard dashboard = getDashboardById(dashboardId);
+        if (!dashboard.isSystem()) {
+            return widgets.stream()
+                          .map(widget -> updateDashboardWidget(dashboardId, widget))
+                          .collect(Collectors.toList());
+        } else {
+            throw new IllegalOperationException(DASHBOARD_CAN_NOT_BE_UPDATED, ERR_MSG_UNEDITABLE_DASHBOARD_CANT_BE_ALTERED);
+        }
     }
 
     @Transactional
@@ -145,47 +154,76 @@ public class DashboardService {
 
     @Transactional
     public Attribute createDashboardAttribute(Long dashboardId, Attribute attribute) {
-        Attribute usedAttribute = retrieveAttributesByDashboardId(dashboardId).stream()
-                                                                        .filter(attr -> attr.getKey().equals(attribute.getKey()))
-                                                                        .findAny()
-                                                                        .orElse(null);
-        if (usedAttribute != null) {
-            String message = String.format(ERR_MSG_DASHBOARD_ATTRIBUTE_ALREADY_IN_USE, usedAttribute.getKey(), dashboardId);
-            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, message);
+        Dashboard dashboard = getDashboardById(dashboardId);
+        if (!dashboard.isSystem()) {
+            Attribute usedAttribute = retrieveAttributesByDashboardId(dashboardId).stream()
+                                                                                  .filter(attr -> attr.getKey().equals(attribute.getKey()))
+                                                                                  .findAny()
+                                                                                  .orElse(null);
+            if (usedAttribute != null) {
+                String message = String.format(ERR_MSG_DASHBOARD_ATTRIBUTE_ALREADY_IN_USE, usedAttribute.getKey(), dashboardId);
+                throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, message);
+            }
+            dashboardMapper.createDashboardAttribute(dashboardId, attribute);
+        } else {
+            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, ERR_MSG_CANNOT_PERSIST_ATTRIBUTE_TO_SYSTEM_DASHBOARD);
         }
-        dashboardMapper.createDashboardAttribute(dashboardId, attribute);
         return attribute;
     }
 
     @Transactional
     public List<Attribute> createDashboardAttributes(Long dashboardId, List<Attribute> attributes) {
-        List<String> dashboardAttributeKeys = retrieveAttributesByDashboardId(dashboardId).stream()
-                                                                                       .map(Attribute::getKey)
-                                                                                       .collect(Collectors.toList());
-        List<String> usedKeys = attributes.stream()
-                                          .filter(attribute -> dashboardAttributeKeys.contains(attribute.getKey()))
-                                          .map(Attribute::getKey)
-                                          .collect(Collectors.toList());
-        if (!usedKeys.isEmpty()) {
-            String keysAsString = String.join(", ", usedKeys);
-            String message = usedKeys.size() > 1 ?
-                    String.format(ERR_MSG_DASHBOARD_ATTRIBUTES_ALREADY_IN_USE, keysAsString, dashboardId) :
-                    String.format(ERR_MSG_DASHBOARD_ATTRIBUTE_ALREADY_IN_USE, usedKeys.get(0), dashboardId);
-            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, message);
+        Dashboard dashboard = getDashboardById(dashboardId);
+        if (!dashboard.isSystem()) {
+            List<String> dashboardAttributeKeys = retrieveAttributesByDashboardId(dashboardId).stream()
+                                                                                              .map(Attribute::getKey)
+                                                                                              .collect(Collectors.toList());
+            List<String> usedKeys = attributes.stream()
+                                              .filter(attribute -> dashboardAttributeKeys.contains(attribute.getKey()))
+                                              .map(Attribute::getKey)
+                                              .collect(Collectors.toList());
+            if (!usedKeys.isEmpty()) {
+                String keysAsString = String.join(", ", usedKeys);
+                String message = usedKeys.size() > 1 ?
+                        String.format(ERR_MSG_DASHBOARD_ATTRIBUTES_ALREADY_IN_USE, keysAsString, dashboardId) :
+                        String.format(ERR_MSG_DASHBOARD_ATTRIBUTE_ALREADY_IN_USE, usedKeys.get(0), dashboardId);
+                throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, message);
+            }
+            dashboardMapper.createDashboardAttributes(dashboardId, attributes);
+        } else {
+            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, ERR_MSG_CANNOT_PERSIST_ATTRIBUTE_TO_SYSTEM_DASHBOARD, dashboardId);
         }
-        dashboardMapper.createDashboardAttributes(dashboardId, attributes);
         return attributes;
+    }
+
+    @Transactional(readOnly = true)
+    public Dashboard retrieveByAttributeId(Long attributeId) {
+        Dashboard dashboard = dashboardMapper.getByAttributeId(attributeId);
+        if (dashboard == null) {
+            throw new ResourceNotFoundException(DASHBOARD_NOT_FOUND, ERR_MSG_DASHBOARD_CAN_NOT_BE_FOUND_BY_ATTRIBUTE_ID, attributeId);
+        }
+        return dashboard;
     }
 
     @Transactional
     public Attribute updateAttribute(Attribute attribute) {
-        dashboardMapper.updateAttribute(attribute);
+        Dashboard dashboard = retrieveByAttributeId(attribute.getId());
+        if (!dashboard.isSystem()) {
+            dashboardMapper.updateAttribute(attribute);
+        } else {
+            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, ERR_MSG_CANNOT_PERSIST_ATTRIBUTE_TO_SYSTEM_DASHBOARD);
+        }
         return attribute;
     }
 
     @Transactional
     public void removeByAttributeById(long attributeId) {
-        dashboardMapper.deleteDashboardAttributeById(attributeId);
+        Dashboard dashboard = retrieveByAttributeId(attributeId);
+        if (!dashboard.isSystem()) {
+            dashboardMapper.deleteDashboardAttributeById(attributeId);
+        } else {
+            throw new IllegalOperationException(DASHBOARD_ATTRIBUTE_CAN_NOT_BE_CREATED, ERR_MSG_CANNOT_PERSIST_ATTRIBUTE_TO_SYSTEM_DASHBOARD);
+        }
     }
 
     @Transactional
